Initial Commit
Problem:
Write a function in C which implements the insertion sort algorithm. we don't have to return the array because it has been passed by reference. Rather, make the return value of insertion sort the number of examinations that the algorithm makes. An examination is defined as a comparison between two elements in the array, regardless of the result of that comparison.